requirements/require2.txt
alabaster==0.7.8
aniso8601==1.2.0
arandr==0.1.9
asn1crypto==0.24.0
automationhat==0.2.0
awscli==1.18.100
AWSIoTPythonSDK==1.4.9
Babel==2.6.0
blinker==1.4
blinkt==0.1.2
boto3==1.17.6
botocore==1.20.6
buttonshim==0.0.2
cachetools==3.1.1
Cap1xxx==0.1.3
certifi==2018.8.24
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
configparser==3.5.0b2
cookies==2.2.1
cryptography==2.8
docutils==0.14
drumhat==0.1.0
entrypoints==0.3
enum34==1.1.6
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==0.12.1
Flask-Ask==0.9.8
fourletterphat==0.1.0
funcsigs==1.0.2
futures==3.3.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-texttospeech==0.5.0
googleapis-common-protos==1.6.0
gpiozero==1.5.1
grpcio==1.26.0
gyp==0.1
httplib2==0.19.0
idna==2.6
imagesize==1.0.0
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.10
jmespath==0.10.0
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.0
mcpi==0.1.1
microdotphat==0.2.1
mock==2.0.0
mote==0.0.4
motephat==0.0.3
numpy==1.16.2
oauthlib==2.1.0
olefile==0.46
packaging==19.0
paho-mqtt==1.5.1
pantilthat==0.0.7
pbr==4.2.0
phatbeat==0.1.1
phue==1.1
pianohat==0.1.0
picamera==1.13
picraft==1.0
piglow==1.2.5
pigpio==1.44
Pillow==5.4.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pyfingerprint==1.5
pyflakes==2.0.0
pygame==1.9.4.post1
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pyOpenSSL==17.0.0
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
rainbowhat==0.1.0
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
roman==2.0.0
rpi-ws281x==1.0.0
RPi.GPIO==0.7.0
rsa==4.0
RTIMULib==7.2.1
s3transfer==0.3.3
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
sense-emu==1.1
sense-hat==2.2.0
simplejson==3.16.0
six==1.15.0
skywriter==0.0.7
sn3218==1.2.7
Sphinx==1.8.4
spidev==3.4
touchphat==0.0.1
twython==3.7.0
typing==3.6.6
unicornhathd==0.0.4
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==0.14.1
